create post component
